requirements.txt
absl-py==0.8.0
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
asyncore-wsgi==0.0.9
autopep8==1.4.4
certifi==2019.9.11
cffi==1.12.3
cftime==1.0.3.4
chardet==3.0.4
colorama==0.4.1
conda==4.7.12
conda-package-handling==1.6.0
ConfigSpace==0.4.10
cryptography==2.7
cycler==0.10.0
flatbuffers==1.11
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
hanging-threads==2.0.5
hpbandster==0.7.4
hyperas==0.4.1
hyperopt==0.1.2
idna==2.8
imbalanced-learn==0.5.0
imblearn==0.0
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==0.9.1
kerasplotlib==0.1.4
kiwisolver==1.1.0
llvmlite==0.29.0
Markdown==3.1.1
matplotlib==3.1.1
netifaces==0.10.9
numba==0.45.1
numpy==1.17.2
opt-einsum==3.0.1
pandas==0.25.1
patsy==0.5.1
protobuf==3.9.2
psutil==5.6.3
psycopg2==2.8.3
py-spy==0.1.11
py4j==0.10.7
pyarrow==0.14.1
pycosat==0.6.3
pycparser==2.19
pyfolio==0.9.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
Pyro4==4.76
PySocks==1.7.1
pyspark==2.4.4
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
retrying==1.3.3
ruamel-yaml==0.15.46
scikit-learn==0.21.3
scikit-optimize==0.5.2
scipy==1.3.1
seaborn==0.9.0
serpent==1.28
setproctitle==1.1.10
six==1.12.0
sklearn==0.0
statsmodels==0.10.1
tabulate==0.8.5
tb-nightly==2.1.0a20190925
tensorboard==2.0.0
tensorflow==2.0.0rc2
tensorflow-estimator-2.0-preview==1.14.0.dev2019061821
termcolor==1.1.0
terminaltables==3.1.0
tf-estimator-nightly==1.14.0.dev2019080601
tf-nightly-2.0-preview==2.0.0.dev20190926
toml==0.10.0
tqdm==4.36.1
urllib3==1.24.2
web-pdb==1.5.1
Werkzeug==0.16.0
wrapt==1.11.2
